requirements.txt
python-telegram-bot==21.9
spotipy==2.24.0
yt-dlp==2024.12.13
requests==2.32.3